add git config entry ditz.central-repo
This parameter allows to easily share one central ditz database in a git
repository among some subrepositories. The README.md contains an example
configuration.

If you configured your repositories as described in the README.md file you can
invoke the `git ditz ...` command anywhere in the central- or subrepository and
the command will be invoked as if you invoked it within the central repository.